+++ src/main/java/com/baisylia/modestmining/screen/MillstoneScreen.java
@@ -2,32 +2,18 @@ package com.baisylia.modestmining.screen;
 
 import com.baisylia.modestmining.ModestMining;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.gui.components.ImageButton;
-import net.minecraft.client.gui.components.WidgetSprites;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
-import net.minecraft.client.gui.screens.recipebook.RecipeBookComponent;
-import net.minecraft.client.gui.screens.recipebook.RecipeUpdateListener;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.inventory.ClickType;
-import net.minecraft.world.inventory.Slot;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class MillstoneScreen extends AbstractContainerScreen<MillstoneMenu> implements RecipeUpdateListener {
+public class MillstoneScreen extends AbstractContainerScreen<MillstoneMenu> {
 
     private static final ResourceLocation TEXTURE =
             ResourceLocation.fromNamespaceAndPath(ModestMining.MOD_ID, "textures/gui/millstone_gui.png");
-
-    private static final WidgetSprites RECIPE_BUTTON_SPRITES = new WidgetSprites(
-            ResourceLocation.withDefaultNamespace("recipe_book/button"),
-            ResourceLocation.withDefaultNamespace("recipe_book/button_highlighted")
-    );
-
-    public final RecipeBookComponent recipeBookComponent = new MillingRecipeBookComponent();
-    private boolean widthTooNarrow;
 
     public MillstoneScreen(MillstoneMenu pMenu, Inventory pPlayerInventory, Component pTitle) {
         super(pMenu, pPlayerInventory, pTitle);
@@ -36,17 +22,7 @@ public class MillstoneScreen extends AbstractContainerScreen<MillstoneMenu> impl
     @Override
     public void init() {
         super.init();
-        this.widthTooNarrow = this.width < 379;
         this.titleLabelX = 28;
-        this.recipeBookComponent.init(this.width, this.height, this.minecraft, this.widthTooNarrow, this.menu);
-        this.leftPos = this.recipeBookComponent.updateScreenPosition(this.width, this.imageWidth);
-        this.addRenderableWidget(new ImageButton(this.leftPos + 5, this.height / 2 - 49, 20, 18, RECIPE_BUTTON_SPRITES, (button) -> {
-            this.recipeBookComponent.toggleVisibility();
-            this.leftPos = this.recipeBookComponent.updateScreenPosition(this.width, this.imageWidth);
-            button.setPosition(this.leftPos + 5, this.height / 2 - 49);
-        }));
-        this.addWidget(this.recipeBookComponent);
-        this.setInitialFocus(this.recipeBookComponent);
     }
 
     @Override
@@ -65,28 +41,10 @@ public class MillstoneScreen extends AbstractContainerScreen<MillstoneMenu> impl
     }
 
     @Override
-    protected void containerTick() {
-        super.containerTick();
-        this.recipeBookComponent.tick();
-    }
-
-    @Override
     public void render(GuiGraphics guiGraphics, int mouseX, int mouseY, float delta) {
         super.render(guiGraphics, mouseX, mouseY, delta);
-
-        if (this.recipeBookComponent.isVisible() && this.widthTooNarrow) {
-            this.renderBg(guiGraphics, delta, mouseX, mouseY);
-            this.recipeBookComponent.render(guiGraphics, mouseX, mouseY, delta);
-        } else {
-            this.recipeBookComponent.render(guiGraphics, mouseX, mouseY, delta);
-            super.render(guiGraphics, mouseX, mouseY, delta);
-            this.recipeBookComponent.renderGhostRecipe(guiGraphics, this.leftPos, this.topPos, true, delta);
-        }
-
-        this.recipeBookComponent.renderTooltip(guiGraphics, this.leftPos, this.topPos, mouseX, mouseY);
         this.renderTooltip(guiGraphics, mouseX, mouseY);
 
-        // Power Tooltip
         if (isHovering(61, 55, 24, 17, mouseX, mouseY)) {
             List<Component> tooltip = new ArrayList<>();
             if (menu.isFueled()) {
@@ -102,41 +60,5 @@ public class MillstoneScreen extends AbstractContainerScreen<MillstoneMenu> impl
     protected void renderLabels(GuiGraphics guiGraphics, int mouseX, int mouseY) {
         super.renderLabels(guiGraphics, mouseX, mouseY);
         guiGraphics.drawString(this.font, this.playerInventoryTitle, 8, (this.imageHeight - 96 + 2), 4210752, false);
-    }
-
-    @Override
-    protected boolean isHovering(int x, int y, int width, int height, double mouseX, double mouseY) {
-        return (!this.widthTooNarrow || !this.recipeBookComponent.isVisible()) && super.isHovering(x, y, width, height, mouseX, mouseY);
-    }
-
-    @Override
-    public boolean mouseClicked(double mouseX, double mouseY, int buttonId) {
-        if (this.recipeBookComponent.mouseClicked(mouseX, mouseY, buttonId)) {
-            this.setFocused(this.recipeBookComponent);
-            return true;
-        }
-        return this.widthTooNarrow && this.recipeBookComponent.isVisible() || super.mouseClicked(mouseX, mouseY, buttonId);
-    }
-
-    @Override
-    protected boolean hasClickedOutside(double mouseX, double mouseY, int x, int y, int buttonIdx) {
-        boolean flag = mouseX < (double) x || mouseY < (double) y || mouseX >= (double) (x + this.imageWidth) || mouseY >= (double) (y + this.imageHeight);
-        return flag && this.recipeBookComponent.hasClickedOutside(mouseX, mouseY, this.leftPos, this.topPos, this.imageWidth, this.imageHeight, buttonIdx);
-    }
-
-    @Override
-    protected void slotClicked(Slot pSlot, int pSlotId, int pMouseButton, ClickType pType) {
-        super.slotClicked(pSlot, pSlotId, pMouseButton, pType);
-        this.recipeBookComponent.slotClicked(pSlot);
-    }
-
-    @Override
-    public void recipesUpdated() {
-        this.recipeBookComponent.recipesUpdated();
-    }
-
-    @Override
-    public RecipeBookComponent getRecipeBookComponent() {
-        return this.recipeBookComponent;
     }
 }