+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - UNRELEASED
+## [1.1.0] - 2026-08-12
+
+### Added
+
+- Added copper screws to archeology loot.
+- Added a config option to disable copper screws in loot.
+
+### Fixed
+
+- Fixed millstone texture.
+- Fixed recipe duplication bugs.
+- Fixed bronze and steel sword recipes.
+- Fixed bronze and steel armor being stackable.
+- Fixed clam model.
+- Fixed clam duplication.
+- Fixed issues with the JEI plugin.
+
+### Removed
+
+- Removed recipe book category (wasn't working properly anyway).
+
+## [1.0.0] - 2026-08-11
 
 - Initial release.